Flexible Violin/Box Plot

This function generates a violin or box plot based on the provided data, parameter, and dose information.

Usage

flexible_violinboxplot(
  boxplotdata,
  parameter,
  xvars,
  colorvars,
  varvalstofilter,
  columns_to_hover,
  box = TRUE,
  plotly = TRUE
)

Arguments

boxplotdata

A list containing the data to be plotted. It should have a data element with dose information and a formatted element with the data to be plotted.

parameter

A string specifying the parameter to be plotted.

xvars

Variables for the x axis.

colorvars

Variables for the color aesthetic.

varvalstofilter

Variables to filter on.

columns_to_hover

A character vector indicating the column names from result_data that should be used to identify when hovering the plotly outputs

box

A logical value indicating whether to plot a box plot (TRUE) or a violin plot (FALSE). Default is TRUE.

plotly

A logical value defining if the output is plotly (TRUE, default) or ggplot otherwise (FALSE)

Value

A plotly object representing the violin or box plot.