Parses the tlg definitions described in yaml file into a list object. Custom YAML parsing is needed for templating - the user can define templates for a definition. The function detects appropriate keywords and copies information from specified templates.

Parses the tlg definitions described in yaml file into a list object. Custom YAML parsing is needed for templating - the user can define templates for a definition. The function detects appropriate keywords and copies information from specified templates.

Usage

parse_tlg_definitions(path = NULL, defs = NULL)

Arguments

path

Path to a YAML file containing TLG definitions.

defs

A list of TLG definitions read from a YAML file.

Value

A full list of available TLG definitions.

Details

If both path and defs are provided, path will be used and definitions will be loaded from the file. If neither is given, as a default, the function will load system file in shiny/tlg.yaml.